release v5.0x1.1
4d6e761 aiq: Update version to v5.0x1.1
3c8842c hwi: Sync kernel driver header to v2.1.0
b7d94ec j2s: parser tool: support both arm and arm64
6d266f8 SmartIr: update smartIr sample
....

+++ include/algos/agic/rk_aiq_types_algo_agic_int.h
@@ -155,8 +155,8 @@ typedef struct AgicConfigV21_s {
     unsigned short regmingradthr2;
     unsigned short regmingradthr1;
     unsigned char gr_ratio;
-    float noise_scale;
-    float noise_base;
+    uint8_t noise_scale;
+    uint16_t noise_base;
     float noiseCurve_0;
     float noiseCurve_1;
     float sigma_y[15];
