+++ include/xcore/base/xcam_common.h
@@ -75,7 +75,7 @@ void xcam_free (void *ptr);
   * return, 0 successfully
   *            else, check errno
   */
-int xcam_device_ioctl (int fd, int cmd, void *arg);
+int xcam_device_ioctl (int fd, unsigned long cmd, void *arg);
 const char *xcam_fourcc_to_string (uint32_t fourcc);
 
 static inline uint32_t
